[L01] Numerous uint types and resultant casts increase system complexity (#88)
* Fix to uint256

* Fix compile warnings

* Fix compile warnings

* Fix compile warnings

* Fix compile warnings

* Fix deposit error

* Nit

* Convert 104 to 256

* Convert 128 to 256

* Remove unnecessary casting

* Fix tests

* [L16] Error-prone variable shadowing (#105)

* Use numShares instead

* Fix compile errors

* Convert all storage to uint256

* Fixing

* Working compile

* Push

+++ contracts/storage/RibbonDeltaVaultStorage.sol
@@ -9,7 +9,7 @@ abstract contract RibbonDeltaVaultStorageV1 {
     // % of funds to be used for weekly option purchase
     uint256 public optionAllocationPct;
     // Delta vault equivalent of lockedAmount
-    uint104 public balanceBeforePremium;
+    uint256 public balanceBeforePremium;
     // User Id of delta vault in latest gnosis auction
     Vault.AuctionSellOrder public auctionSellOrder;
 }
